refactor: move validate to expose it as receivers (#2419)
## Description

Moves `validate` behavior out of `internal` and onto receivers to their
given types. Since validation requires no external logic, this fits
better and has a cleaner experience. Additionally, validation can be
called against each type itself individually instead of only living at
the package level.

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://github.com/defenseunicorns/zarf/blob/main/CONTRIBUTING.md#developer-workflow)
followed

---------
