+++ src/pkg/packager/create.go
@@ -10,7 +10,6 @@ import (
 
 	"github.com/defenseunicorns/pkg/helpers"
 	"github.com/defenseunicorns/zarf/src/config"
-	"github.com/defenseunicorns/zarf/src/internal/packager/validate"
 	"github.com/defenseunicorns/zarf/src/pkg/layout"
 	"github.com/defenseunicorns/zarf/src/pkg/message"
 	"github.com/defenseunicorns/zarf/src/pkg/packager/creator"
@@ -41,7 +40,7 @@ func (p *Packager) Create() (err error) {
 	}
 
 	// Perform early package validation.
-	if err := validate.Run(p.cfg.Pkg); err != nil {
+	if err := p.cfg.Pkg.Validate(); err != nil {
 		return fmt.Errorf("unable to validate package: %w", err)
 	}
