+++ src/pkg/packager/dev.go
@@ -11,7 +11,6 @@ import (
 
 	"github.com/defenseunicorns/pkg/helpers"
 	"github.com/defenseunicorns/zarf/src/config"
-	"github.com/defenseunicorns/zarf/src/internal/packager/validate"
 	"github.com/defenseunicorns/zarf/src/pkg/layout"
 	"github.com/defenseunicorns/zarf/src/pkg/message"
 	"github.com/defenseunicorns/zarf/src/pkg/packager/creator"
@@ -53,7 +52,7 @@ func (p *Packager) DevDeploy() error {
 		return err
 	}
 
-	if err := validate.Run(p.cfg.Pkg); err != nil {
+	if err := p.cfg.Pkg.Validate(); err != nil {
 		return fmt.Errorf("unable to validate package: %w", err)
 	}
